fastuidraw/glsl: avoid implicit int to uint casting

+++ src/fastuidraw/glsl/private/backend_shaders.cpp
@@ -278,54 +278,54 @@ ShaderSetCreatorStrokingConstants(void)
 
   m_stroke_constants
     /* offset types of StrokedPoint */
-    .add_macro("fastuidraw_stroke_offset_sub_edge", StrokedPoint::offset_sub_edge)
-    .add_macro("fastuidraw_stroke_offset_shared_with_edge", StrokedPoint::offset_shared_with_edge)
-    .add_macro("fastuidraw_stroke_offset_rounded_join", StrokedPoint::offset_rounded_join)
+    .add_macro("fastuidraw_stroke_offset_sub_edge", uint32_t(StrokedPoint::offset_sub_edge))
+    .add_macro("fastuidraw_stroke_offset_shared_with_edge", uint32_t(StrokedPoint::offset_shared_with_edge))
+    .add_macro("fastuidraw_stroke_offset_rounded_join", uint32_t(StrokedPoint::offset_rounded_join))
 
-    .add_macro("fastuidraw_stroke_offset_miter_bevel_join", StrokedPoint::offset_miter_bevel_join)
-    .add_macro("fastuidraw_stroke_offset_miter_join", StrokedPoint::offset_miter_join)
-    .add_macro("fastuidraw_stroke_offset_miter_clip_join", StrokedPoint::offset_miter_clip_join)
+    .add_macro("fastuidraw_stroke_offset_miter_bevel_join", uint32_t(StrokedPoint::offset_miter_bevel_join))
+    .add_macro("fastuidraw_stroke_offset_miter_join", uint32_t(StrokedPoint::offset_miter_join))
+    .add_macro("fastuidraw_stroke_offset_miter_clip_join", uint32_t(StrokedPoint::offset_miter_clip_join))
 
-    .add_macro("fastuidraw_stroke_offset_rounded_cap", StrokedPoint::offset_rounded_cap)
-    .add_macro("fastuidraw_stroke_offset_square_cap", StrokedPoint::offset_square_cap)
-    .add_macro("fastuidraw_stroke_offset_adjustable_cap", StrokedPoint::offset_adjustable_cap)
+    .add_macro("fastuidraw_stroke_offset_rounded_cap", uint32_t(StrokedPoint::offset_rounded_cap))
+    .add_macro("fastuidraw_stroke_offset_square_cap", uint32_t(StrokedPoint::offset_square_cap))
+    .add_macro("fastuidraw_stroke_offset_adjustable_cap", uint32_t(StrokedPoint::offset_adjustable_cap))
 
     /* bit masks for StrokedPoint::m_packed_data */
-    .add_macro("fastuidraw_stroke_sin_sign_mask", StrokedPoint::sin_sign_mask)
-    .add_macro("fastuidraw_stroke_normal0_y_sign_mask", StrokedPoint::normal0_y_sign_mask)
-    .add_macro("fastuidraw_stroke_normal1_y_sign_mask", StrokedPoint::normal1_y_sign_mask)
-    .add_macro("fastuidraw_stroke_lambda_negated_mask", StrokedPoint::lambda_negated_mask)
-    .add_macro("fastuidraw_stroke_boundary_bit", StrokedPoint::boundary_bit)
-    .add_macro("fastuidraw_stroke_join_mask", StrokedPoint::join_mask)
-    .add_macro("fastuidraw_stroke_bevel_edge_mask", StrokedPoint::bevel_edge_mask)
-    .add_macro("fastuidraw_stroke_end_sub_edge_mask", StrokedPoint::end_sub_edge_mask)
-    .add_macro("fastuidraw_stroke_adjustable_cap_ending_mask", StrokedPoint::adjustable_cap_ending_mask)
-    .add_macro("fastuidraw_stroke_adjustable_cap_end_contour_mask", StrokedPoint::adjustable_cap_is_end_contour_mask)
-    .add_macro("fastuidraw_stroke_depth_bit0", StrokedPoint::depth_bit0)
-    .add_macro("fastuidraw_stroke_depth_num_bits", StrokedPoint::depth_num_bits)
-    .add_macro("fastuidraw_stroke_offset_type_bit0", StrokedPoint::offset_type_bit0)
-    .add_macro("fastuidraw_stroke_offset_type_num_bits", StrokedPoint::offset_type_num_bits);
+    .add_macro("fastuidraw_stroke_sin_sign_mask", uint32_t(StrokedPoint::sin_sign_mask))
+    .add_macro("fastuidraw_stroke_normal0_y_sign_mask", uint32_t(StrokedPoint::normal0_y_sign_mask))
+    .add_macro("fastuidraw_stroke_normal1_y_sign_mask", uint32_t(StrokedPoint::normal1_y_sign_mask))
+    .add_macro("fastuidraw_stroke_lambda_negated_mask", uint32_t(StrokedPoint::lambda_negated_mask))
+    .add_macro("fastuidraw_stroke_boundary_bit", uint32_t(StrokedPoint::boundary_bit))
+    .add_macro("fastuidraw_stroke_join_mask", uint32_t(StrokedPoint::join_mask))
+    .add_macro("fastuidraw_stroke_bevel_edge_mask", uint32_t(StrokedPoint::bevel_edge_mask))
+    .add_macro("fastuidraw_stroke_end_sub_edge_mask", uint32_t(StrokedPoint::end_sub_edge_mask))
+    .add_macro("fastuidraw_stroke_adjustable_cap_ending_mask", uint32_t(StrokedPoint::adjustable_cap_ending_mask))
+    .add_macro("fastuidraw_stroke_adjustable_cap_end_contour_mask", uint32_t(StrokedPoint::adjustable_cap_is_end_contour_mask))
+    .add_macro("fastuidraw_stroke_depth_bit0", uint32_t(StrokedPoint::depth_bit0))
+    .add_macro("fastuidraw_stroke_depth_num_bits", uint32_t(StrokedPoint::depth_num_bits))
+    .add_macro("fastuidraw_stroke_offset_type_bit0", uint32_t(StrokedPoint::offset_type_bit0))
+    .add_macro("fastuidraw_stroke_offset_type_num_bits", uint32_t(StrokedPoint::offset_type_num_bits));
 
   m_arc_stroke_constants
     /* offset types of ArcStrokedPoint */
-    .add_macro("fastuidraw_arc_stroke_arc_point", ArcStrokedPoint::offset_arc_point)
-    .add_macro("fastuidraw_arc_stroke_line_segment", ArcStrokedPoint::offset_line_segment)
-    .add_macro("fastuidraw_arc_stroke_dashed_capper", ArcStrokedPoint::offset_arc_point_dashed_capper)
+    .add_macro("fastuidraw_arc_stroke_arc_point", uint32_t(ArcStrokedPoint::offset_arc_point))
+    .add_macro("fastuidraw_arc_stroke_line_segment", uint32_t(ArcStrokedPoint::offset_line_segment))
+    .add_macro("fastuidraw_arc_stroke_dashed_capper", uint32_t(ArcStrokedPoint::offset_arc_point_dashed_capper))
 
     /* bit masks for ArcStrokedPoint::m_packed_data */
-    .add_macro("fastuidraw_arc_stroke_extend_mask", ArcStrokedPoint::extend_mask)
-    .add_macro("fastuidraw_arc_stroke_join_mask", ArcStrokedPoint::join_mask)
-    .add_macro("fastuidraw_arc_stroke_distance_constant_on_primitive_mask", ArcStrokedPoint::distance_constant_on_primitive_mask)
-    .add_macro("fastuidraw_arc_stroke_beyond_boundary_mask", ArcStrokedPoint::beyond_boundary_mask)
-    .add_macro("fastuidraw_arc_stroke_inner_stroking_mask", ArcStrokedPoint::inner_stroking_mask)
-    .add_macro("fastuidraw_arc_stroke_move_to_arc_center_mask", ArcStrokedPoint::move_to_arc_center_mask)
-    .add_macro("fastuidraw_arc_stroke_end_segment_mask", ArcStrokedPoint::end_segment_mask)
-    .add_macro("fastuidraw_arc_stroke_boundary_bit", ArcStrokedPoint::boundary_bit)
-    .add_macro("fastuidraw_arc_stroke_boundary_mask", ArcStrokedPoint::boundary_mask)
-    .add_macro("fastuidraw_arc_stroke_depth_bit0", ArcStrokedPoint::depth_bit0)
-    .add_macro("fastuidraw_arc_stroke_depth_num_bits", ArcStrokedPoint::depth_num_bits)
-    .add_macro("fastuidraw_arc_stroke_offset_type_bit0", ArcStrokedPoint::offset_type_bit0)
-    .add_macro("fastuidraw_arc_stroke_offset_type_num_bits", ArcStrokedPoint::offset_type_num_bits);
+    .add_macro("fastuidraw_arc_stroke_extend_mask", uint32_t(ArcStrokedPoint::extend_mask))
+    .add_macro("fastuidraw_arc_stroke_join_mask", uint32_t(ArcStrokedPoint::join_mask))
+    .add_macro("fastuidraw_arc_stroke_distance_constant_on_primitive_mask", uint32_t(ArcStrokedPoint::distance_constant_on_primitive_mask))
+    .add_macro("fastuidraw_arc_stroke_beyond_boundary_mask", uint32_t(ArcStrokedPoint::beyond_boundary_mask))
+    .add_macro("fastuidraw_arc_stroke_inner_stroking_mask", uint32_t(ArcStrokedPoint::inner_stroking_mask))
+    .add_macro("fastuidraw_arc_stroke_move_to_arc_center_mask", uint32_t(ArcStrokedPoint::move_to_arc_center_mask))
+    .add_macro("fastuidraw_arc_stroke_end_segment_mask", uint32_t(ArcStrokedPoint::end_segment_mask))
+    .add_macro("fastuidraw_arc_stroke_boundary_bit", uint32_t(ArcStrokedPoint::boundary_bit))
+    .add_macro("fastuidraw_arc_stroke_boundary_mask", uint32_t(ArcStrokedPoint::boundary_mask))
+    .add_macro("fastuidraw_arc_stroke_depth_bit0", uint32_t(ArcStrokedPoint::depth_bit0))
+    .add_macro("fastuidraw_arc_stroke_depth_num_bits", uint32_t(ArcStrokedPoint::depth_num_bits))
+    .add_macro("fastuidraw_arc_stroke_offset_type_bit0", uint32_t(ArcStrokedPoint::offset_type_bit0))
+    .add_macro("fastuidraw_arc_stroke_offset_type_num_bits", uint32_t(ArcStrokedPoint::offset_type_num_bits));
 }
 
 void
@@ -343,19 +343,19 @@ create_macro_set(ShaderSource::MacroSet &dst, bool render_pass_varies) const
   else
     {
       dst
-        .add_macro("fastuidraw_stroke_sub_shader_render_pass_bit0", m_stroke_render_pass_bit0)
-        .add_macro("fastuidraw_stroke_sub_shader_render_pass_num_bits", stroke_render_pass_num_bits);
+        .add_macro("fastuidraw_stroke_sub_shader_render_pass_bit0", uint32_t(m_stroke_render_pass_bit0))
+        .add_macro("fastuidraw_stroke_sub_shader_render_pass_num_bits", uint32_t(stroke_render_pass_num_bits));
     }
 
   dst
-    .add_macro("fastuidraw_stroke_sub_shader_dash_style_bit0", stroke_dash_style_bit0)
-    .add_macro("fastuidraw_stroke_sub_shader_dash_style_num_bits", m_stroke_dash_style_num_bits)
-    .add_macro("fastuidraw_stroke_dashed_flat_caps", PainterEnums::flat_caps)
-    .add_macro("fastuidraw_stroke_dashed_rounded_caps", PainterEnums::rounded_caps)
-    .add_macro("fastuidraw_stroke_dashed_square_caps", PainterEnums::square_caps)
-    .add_macro("fastuidraw_stroke_aa_pass1", uber_stroke_aa_pass1)
-    .add_macro("fastuidraw_stroke_aa_pass2", uber_stroke_aa_pass2)
-    .add_macro("fastuidraw_stroke_non_aa", uber_stroke_non_aa);
+    .add_macro("fastuidraw_stroke_sub_shader_dash_style_bit0", uint32_t(stroke_dash_style_bit0))
+    .add_macro("fastuidraw_stroke_sub_shader_dash_style_num_bits", uint32_t(m_stroke_dash_style_num_bits))
+    .add_macro("fastuidraw_stroke_dashed_flat_caps", uint32_t(PainterEnums::flat_caps))
+    .add_macro("fastuidraw_stroke_dashed_rounded_caps", uint32_t(PainterEnums::rounded_caps))
+    .add_macro("fastuidraw_stroke_dashed_square_caps", uint32_t(PainterEnums::square_caps))
+    .add_macro("fastuidraw_stroke_aa_pass1", uint32_t(uber_stroke_aa_pass1))
+    .add_macro("fastuidraw_stroke_aa_pass2", uint32_t(uber_stroke_aa_pass2))
+    .add_macro("fastuidraw_stroke_non_aa", uint32_t(uber_stroke_non_aa));
 }
 
 //////////////////////////////////////////
@@ -405,46 +405,46 @@ ShaderSetCreator(enum PainterCompositeShader::shader_type composite_tp,
     }
 
   m_common_glyph_attribute_macros
-    .add_macro("FASTUIDRAW_GLYPH_TEXEL_X_BIT0", PainterAttributeDataFillerGlyphs::bit0_x_texel)
-    .add_macro("FASTUIDRAW_GLYPH_TEXEL_Y_BIT0", PainterAttributeDataFillerGlyphs::bit0_y_texel)
-    .add_macro("FASTUIDRAW_GLYPH_TEXEL_Z_BIT0", PainterAttributeDataFillerGlyphs::bit0_z_texel)
-    .add_macro("FASTUIDRAW_GLYPH_TEXEL_X_NUMBITS", PainterAttributeDataFillerGlyphs::num_texel_coord_bits)
-    .add_macro("FASTUIDRAW_GLYPH_TEXEL_Y_NUMBITS", PainterAttributeDataFillerGlyphs::num_texel_coord_bits)
-    .add_macro("FASTUIDRAW_GLYPH_TEXEL_Z_NUMBITS", PainterAttributeDataFillerGlyphs::num_texel_coord_bits)
-    .add_macro("FASTUIDRAW_GLYPH_TEXEL_INVALID_MASK", PainterAttributeDataFillerGlyphs::invalid_mask);
+    .add_macro("FASTUIDRAW_GLYPH_TEXEL_X_BIT0", uint32_t(PainterAttributeDataFillerGlyphs::bit0_x_texel))
+    .add_macro("FASTUIDRAW_GLYPH_TEXEL_Y_BIT0", uint32_t(PainterAttributeDataFillerGlyphs::bit0_y_texel))
+    .add_macro("FASTUIDRAW_GLYPH_TEXEL_Z_BIT0", uint32_t(PainterAttributeDataFillerGlyphs::bit0_z_texel))
+    .add_macro("FASTUIDRAW_GLYPH_TEXEL_X_NUMBITS", uint32_t(PainterAttributeDataFillerGlyphs::num_texel_coord_bits))
+    .add_macro("FASTUIDRAW_GLYPH_TEXEL_Y_NUMBITS", uint32_t(PainterAttributeDataFillerGlyphs::num_texel_coord_bits))
+    .add_macro("FASTUIDRAW_GLYPH_TEXEL_Z_NUMBITS", uint32_t(PainterAttributeDataFillerGlyphs::num_texel_coord_bits))
+    .add_macro("FASTUIDRAW_GLYPH_TEXEL_INVALID_MASK", uint32_t(PainterAttributeDataFillerGlyphs::invalid_mask));
 
   m_glyph_restricted_rays_macros
-    .add_macro("HIERARCHY_NODE_BIT", GlyphRenderDataRestrictedRays::hierarchy_is_node_bit)
-    .add_macro("HIERARCHY_NODE_MASK", FASTUIDRAW_MASK(GlyphRenderDataRestrictedRays::hierarchy_is_node_bit, 1u))
-    .add_macro("HIERARCHY_SPLIT_COORD_BIT", GlyphRenderDataRestrictedRays::hierarchy_splitting_coordinate_bit)
-    .add_macro("HIERARCHY_CHILD0_BIT", GlyphRenderDataRestrictedRays::hierarchy_child0_offset_bit0)
-    .add_macro("HIERARCHY_CHILD1_BIT", GlyphRenderDataRestrictedRays::hierarchy_child1_offset_bit0)
-    .add_macro("HIERARCHY_CHILD_NUM_BITS", GlyphRenderDataRestrictedRays::hierarchy_child_offset_numbits)
-    .add_macro("HIERARCHY_CURVE_LIST_BIT0", GlyphRenderDataRestrictedRays::hierarchy_leaf_curve_list_bit0)
-    .add_macro("HIERARCHY_CURVE_LIST_NUM_BITS", GlyphRenderDataRestrictedRays::hierarchy_leaf_curve_list_numbits)
-    .add_macro("HIERARCHY_CURVE_LIST_SIZE_BIT0", GlyphRenderDataRestrictedRays::hierarchy_leaf_curve_list_size_bit0)
-    .add_macro("HIERARCHY_CURVE_LIST_SIZE_NUM_BITS", GlyphRenderDataRestrictedRays::hierarchy_leaf_curve_list_size_numbits)
-    .add_macro("POINT_COORDINATE_BIAS", GlyphRenderDataRestrictedRays::point_coordinate_bias)
-    .add_macro("POINT_COORDINATE_NUM_BITS", GlyphRenderDataRestrictedRays::point_coordinate_numbits)
-    .add_macro("POINT_X_COORDINATE_BIT0", GlyphRenderDataRestrictedRays::point_x_coordinate_bit0)
-    .add_macro("POINT_Y_COORDINATE_BIT0", GlyphRenderDataRestrictedRays::point_y_coordinate_bit0)
-    .add_macro("RIGHT_CORNER_MASK", GlyphAttribute::right_corner_mask)
-    .add_macro("TOP_CORNER_MASK", GlyphAttribute::top_corner_mask)
-    .add_macro("WINDING_VALUE_BIAS", GlyphRenderDataRestrictedRays::winding_bias)
-    .add_macro("WINDING_VALUE_BIT0", GlyphRenderDataRestrictedRays::winding_value_bit0)
-    .add_macro("WINDING_VALUE_NUM_BITS", GlyphRenderDataRestrictedRays::winding_value_numbits)
-    .add_macro("POSITION_DELTA_BIAS", GlyphRenderDataRestrictedRays::delta_bias)
-    .add_macro("POSITION_DELTA_DIVIDE", GlyphRenderDataRestrictedRays::delta_div_factor)
-    .add_macro("POSITION_DELTA_X_BIT0", GlyphRenderDataRestrictedRays::delta_x_bit0)
-    .add_macro("POSITION_DELTA_Y_BIT0", GlyphRenderDataRestrictedRays::delta_y_bit0)
-    .add_macro("POSITION_DELTA_NUM_BITS", GlyphRenderDataRestrictedRays::delta_numbits)
-    .add_macro("CURVE_ENTRY_NUM_BITS", GlyphRenderDataRestrictedRays::curve_numbits)
-    .add_macro("CURVE_ENTRY0_BIT0", GlyphRenderDataRestrictedRays::curve_entry0_bit0)
-    .add_macro("CURVE_ENTRY1_BIT0", GlyphRenderDataRestrictedRays::curve_entry1_bit0)
-    .add_macro("CURVE_IS_QUADRATIC_BIT", GlyphRenderDataRestrictedRays::curve_is_quadratic_bit)
-    .add_macro("CURVE_IS_QUADRATIC_MASK", FASTUIDRAW_MASK(GlyphRenderDataRestrictedRays::curve_is_quadratic_bit, 1u))
-    .add_macro("CURVE_BIT0", GlyphRenderDataRestrictedRays::curve_location_bit0)
-    .add_macro("CURVE_NUM_BITS", GlyphRenderDataRestrictedRays::curve_location_numbits);
+    .add_macro("HIERARCHY_NODE_BIT", uint32_t(GlyphRenderDataRestrictedRays::hierarchy_is_node_bit))
+    .add_macro("HIERARCHY_NODE_MASK", uint32_t(FASTUIDRAW_MASK(GlyphRenderDataRestrictedRays::hierarchy_is_node_bit, 1u)))
+    .add_macro("HIERARCHY_SPLIT_COORD_BIT", uint32_t(GlyphRenderDataRestrictedRays::hierarchy_splitting_coordinate_bit))
+    .add_macro("HIERARCHY_CHILD0_BIT", uint32_t(GlyphRenderDataRestrictedRays::hierarchy_child0_offset_bit0))
+    .add_macro("HIERARCHY_CHILD1_BIT", uint32_t(GlyphRenderDataRestrictedRays::hierarchy_child1_offset_bit0))
+    .add_macro("HIERARCHY_CHILD_NUM_BITS", uint32_t(GlyphRenderDataRestrictedRays::hierarchy_child_offset_numbits))
+    .add_macro("HIERARCHY_CURVE_LIST_BIT0", uint32_t(GlyphRenderDataRestrictedRays::hierarchy_leaf_curve_list_bit0))
+    .add_macro("HIERARCHY_CURVE_LIST_NUM_BITS", uint32_t(GlyphRenderDataRestrictedRays::hierarchy_leaf_curve_list_numbits))
+    .add_macro("HIERARCHY_CURVE_LIST_SIZE_BIT0", uint32_t(GlyphRenderDataRestrictedRays::hierarchy_leaf_curve_list_size_bit0))
+    .add_macro("HIERARCHY_CURVE_LIST_SIZE_NUM_BITS", uint32_t(GlyphRenderDataRestrictedRays::hierarchy_leaf_curve_list_size_numbits))
+    .add_macro("POINT_COORDINATE_BIAS", uint32_t(GlyphRenderDataRestrictedRays::point_coordinate_bias))
+    .add_macro("POINT_COORDINATE_NUM_BITS", uint32_t(GlyphRenderDataRestrictedRays::point_coordinate_numbits))
+    .add_macro("POINT_X_COORDINATE_BIT0", uint32_t(GlyphRenderDataRestrictedRays::point_x_coordinate_bit0))
+    .add_macro("POINT_Y_COORDINATE_BIT0", uint32_t(GlyphRenderDataRestrictedRays::point_y_coordinate_bit0))
+    .add_macro("RIGHT_CORNER_MASK", uint32_t(GlyphAttribute::right_corner_mask))
+    .add_macro("TOP_CORNER_MASK", uint32_t(GlyphAttribute::top_corner_mask))
+    .add_macro("WINDING_VALUE_BIAS", uint32_t(GlyphRenderDataRestrictedRays::winding_bias))
+    .add_macro("WINDING_VALUE_BIT0", uint32_t(GlyphRenderDataRestrictedRays::winding_value_bit0))
+    .add_macro("WINDING_VALUE_NUM_BITS", uint32_t(GlyphRenderDataRestrictedRays::winding_value_numbits))
+    .add_macro("POSITION_DELTA_BIAS", uint32_t(GlyphRenderDataRestrictedRays::delta_bias))
+    .add_macro("POSITION_DELTA_DIVIDE", uint32_t(GlyphRenderDataRestrictedRays::delta_div_factor))
+    .add_macro("POSITION_DELTA_X_BIT0", uint32_t(GlyphRenderDataRestrictedRays::delta_x_bit0))
+    .add_macro("POSITION_DELTA_Y_BIT0", uint32_t(GlyphRenderDataRestrictedRays::delta_y_bit0))
+    .add_macro("POSITION_DELTA_NUM_BITS", uint32_t(GlyphRenderDataRestrictedRays::delta_numbits))
+    .add_macro("CURVE_ENTRY_NUM_BITS", uint32_t(GlyphRenderDataRestrictedRays::curve_numbits))
+    .add_macro("CURVE_ENTRY0_BIT0", uint32_t(GlyphRenderDataRestrictedRays::curve_entry0_bit0))
+    .add_macro("CURVE_ENTRY1_BIT0", uint32_t(GlyphRenderDataRestrictedRays::curve_entry1_bit0))
+    .add_macro("CURVE_IS_QUADRATIC_BIT", uint32_t(GlyphRenderDataRestrictedRays::curve_is_quadratic_bit))
+    .add_macro("CURVE_IS_QUADRATIC_MASK", uint32_t(FASTUIDRAW_MASK(GlyphRenderDataRestrictedRays::curve_is_quadratic_bit, 1u)))
+    .add_macro("CURVE_BIT0", uint32_t(GlyphRenderDataRestrictedRays::curve_location_bit0))
+    .add_macro("CURVE_NUM_BITS", uint32_t(GlyphRenderDataRestrictedRays::curve_location_numbits));
 }
 
 varying_list
